Removed 100m and 20m specific HHM processes from the BlueTopoS3Engine
